Move checkpoint to `change` module
To be consistent with the Go version
https://github.com/yorkie-team/yorkie/commit/57ca23a669660ce8191d4cd0d469e48d523c08b2
