+++ src/document/change/checkpoint.ts
